dasda

+++ src/Server/ConnEmails.java
@@ -1,6 +1,6 @@
 package Server;
 
-//import Comunication.Mail;
+
 
 import Comunication.Mail;
 import Comunication.Message;
@@ -57,38 +57,47 @@ public class ConnEmails {
 
     }
 
-    public Message getListMess(String account) throws SQLException {
+    public Message getListMess(String account, String password) throws SQLException {
         try {
-            ArrayList<Message> messages = new ArrayList<>();
-            String sql = "SELECT * FROM {0}";
-            preparedStatement = connection.prepareStatement(MessageFormat.format(sql, string_converter(account)));
-            ResultSet resultSet = preparedStatement.executeQuery();
-            while (resultSet.next()) {
-                messages.add(new Mail("SendMails", account, true, resultSet.getString("sender"),
-                        resultSet.getInt("messId"), resultSet.getString("title"),
-                        resultSet.getString("date_"), resultSet.getString("message")));
-            }
-            if (messages.size() != 0)
-                return new SendMaills("SendMails", null, true, messages);
-            else
-                return new SendMaills("SendMails", null, false, null);
+            ConnAcc connAcc = new ConnAcc();
+            if (connAcc.isAcc(account, password)) {
+                ArrayList<Message> messages = new ArrayList<>();
+                String sql = "SELECT * FROM {0}";
+                preparedStatement = connection.prepareStatement(MessageFormat.format(sql, string_converter(account)));
+                ResultSet resultSet = preparedStatement.executeQuery();
+                while (resultSet.next()) {
+                    messages.add(new Mail("SendMails", account, null,true, resultSet.getString("sender"),
+                            resultSet.getInt("messId"), resultSet.getString("title"),
+                            resultSet.getString("date_"), resultSet.getString("message")));
+                }
+                if (messages.size() != 0)
+                    return new SendMaills("SendMails", account, null, true,messages);
+                else
+                    return new SendMaills("SendMails", account, null, true, null);
+            }else
+                return new SendMaills("Fail", account, null, false, null);
         } catch (Exception e) {
             System.out.println("Fatal error");
-            return new SendMaills("Fail", null, false, null);
+            return new SendMaills("Fail", account, null, false, null);
         }
     }
 
 
-    public boolean deleteMess(String account, int id) throws SQLException {
+    public boolean deleteMess(String account, String password,int id) throws SQLException {
         try {
-            String sql = "DELETE FROM {0} WHERE messId = ?";
-            preparedStatement = connection.prepareStatement(MessageFormat.format(sql, string_converter(account)));
-            preparedStatement.setInt(1, id);
-            if (preparedStatement.executeUpdate() != 0) {
-                System.out.println("Udane usuniecie maila");
-                return true;
-            } else {
-                System.out.println("Brak rekordu lub tabeli");
+            ConnAcc connAcc = new ConnAcc();
+            if (connAcc.isAcc(account, password)) {
+                String sql = "DELETE FROM {0} WHERE messId = ?";
+                preparedStatement = connection.prepareStatement(MessageFormat.format(sql, string_converter(account)));
+                preparedStatement.setInt(1, id);
+                if (preparedStatement.executeUpdate() != 0) {
+                    System.out.println("Udane usuniecie maila");
+                    return true;
+                } else {
+                    System.out.println("Brak rekordu lub tabeli");
+                    return false;
+                }
+            }else {
                 return false;
             }
         } catch (Exception e) {
